requirements.txt
edge-tts==6.1.9
openai==1.3.9
pygame==2.5.2
python-dotenv==1.0.0
elevenlabs==0.2.24
